Move command classes to private API
