+++ source/MaterialXRuntime/Private/PvtCommandEngine.h
@@ -7,18 +7,72 @@
 #define MATERIALX_PVTCOMMANDENGINE_H
 
 
-#include <MaterialXRuntime/RtCommandEngine.h>
+#include <MaterialXRuntime/RtCommand.h>
 
 #include <vector>
 
 namespace MaterialX
 {
 
+/// @class PvtCommand
+/// Base class for runtime commands.
+class PvtCommand
+{
+public:
+    virtual ~PvtCommand() {};
+
+    PvtCommand(const PvtCommand&) = delete;
+    PvtCommand& operator=(const PvtCommand&) = delete;
+
+    /// Execute the command.
+    virtual void execute(RtCommandResult& result) = 0;
+
+    /// Undo the command.
+    virtual void undo(RtCommandResult& result) = 0;
+
+    /// Redo the command.
+    virtual void redo(RtCommandResult& result)
+    {
+        execute(result);
+    }
+
+protected:
+    PvtCommand() {}
+};
+
+/// A shared pointer to a runtime command.
+using PvtCommandPtr = RtSharedPtr<PvtCommand>;
+
+/// @class PvtCommandList
+/// Class for executing lists of multiple commands.
+class PvtCommandList : public PvtCommand
+{
+public:
+    /// Add a command to the batch.
+    void addCommand(PvtCommandPtr cmd);
+
+    /// Clear all commands in the batch.
+    void clearCommands();
+
+    /// Execute the command.
+    void execute(RtCommandResult& result) override;
+
+    /// Undo the command.
+    void undo(RtCommandResult& result)  override;
+
+    /// Redo the command.
+    void redo(RtCommandResult& result)  override;
+
+protected:
+    vector<PvtCommandPtr> _commands;
+};
+
+
 class PvtCommandEngine
 {
 public:
     /// Execute a new command.
-    void execute(RtCommandPtr cmd, RtCommandResult& result);
+    void execute(PvtCommandPtr cmd, RtCommandResult& result);
 
     /// Undo the last previously executed command.
     void undo(RtCommandResult& result);
@@ -31,8 +85,8 @@ public:
     void flushUndoQueue();
 
 private:
-    vector<RtCommandPtr> _undoQueue;
-    vector<RtCommandPtr> _redoQueue;
+    vector<PvtCommandPtr> _undoQueue;
+    vector<PvtCommandPtr> _redoQueue;
 };
 
 }
